+++ src/fpath_extended.ml
@@ -1,9 +1,9 @@
 module Fpath = struct
   include Fpath
   include Fpath0
+  include Path.Export
 end
 
-module Classified_path = Path.Classified_path
 module Absolute_path = Path.Absolute_path
 module Relative_path = Path.Relative_path
 module File_name = File_name