+++ src/path.mli
@@ -31,6 +31,13 @@ module Absolute_path : sig
   val parent : t -> t option
   val chop_prefix : prefix:t -> t -> relative_path Or_error.t
   val chop_suffix : t -> suffix:relative_path -> t Or_error.t
+  val is_dir_path : t -> bool
+  val to_dir_path : t -> t
+
+  (** Converts a Path.t to an Absolute_path.t:
+      - If the path is already absolute, that's the answer.
+      - If the path is relative, it is made absolute by appending it to [root]. *)
+  val relativize : root:t -> Fpath.t -> t
 end
 
 module Relative_path : sig
@@ -42,7 +49,9 @@ module Relative_path : sig
   val to_string : t -> string
 
   (** [of_fpath p] returns a normalized of [p] classified as a relative path.
-      Returns an error if [p] is not a relative path. *)
+      Returns an error if [p] is not a relative path. Note, in particular that
+      due to normalization, ["."] immediately becomes ["./"] (the empty
+      relative path). *)
   val of_fpath : Fpath.t -> t Or_error.t
 
   (** This is a convenient wrapper to compose {!Fpath.of_string} and {!of_fpath}. *)
@@ -51,11 +60,8 @@ module Relative_path : sig
   (** [v str] is [of_string str |> Or_error.ok_exn]. *)
   val v : string -> t
 
-  (** The empty relative path ["."]. *)
-  val dot : t
-
-  (** The current directory ["./"]. *)
-  val dot_slash : t
+  (** The empty relative path ["./"]. *)
+  val empty : t
 
   val append : t -> t -> t
   val extend : t -> File_name.t -> t
@@ -63,43 +69,14 @@ module Relative_path : sig
   val of_list : File_name.t list -> t
   val chop_prefix : prefix:t -> t -> t Or_error.t
   val chop_suffix : t -> suffix:t -> t Or_error.t
+  val is_dir_path : t -> bool
+  val to_dir_path : t -> t
 end
 
-module Classified_path : sig
-  type t =
-    | Absolute of Absolute_path.t
-    | Relative of Relative_path.t
-  [@@deriving compare, equal, hash, sexp_of]
-
-  include Comparable.S with type t := t
-
-  (** [of_fpath p] returns a normalized and classified version of [p]. *)
-  val of_fpath : Fpath.t -> t
-
-  val to_fpath : t -> Fpath.t
-  val to_string : t -> string
-
-  (** This is a convenient wrapper to compose {!Fpath.of_string} and {!of_fpath}. *)
-  val of_string : string -> t Or_error.t
-
-  (** [v str] is [of_string str |> Or_error.ok_exn]. *)
-  val v : string -> t
-
-  (** [Absolute Absolute_path.root], aka ["/"]. *)
-  val root : t
-
-  (** [Relative Relative_path.dot], aka ["."]. *)
-  val dot : t
-
-  (** Relocate p2, if it is relative, to p1. If p2 is absolute, just ignore p1 and
-      return p2. *)
-  val append : t -> t -> t
-
-  val extend : t -> File_name.t -> t
-  val parent : t -> t option
-
-  (** Converts a Path.t to an Absolute_path.t:
-      - If the path is already absolute, it is returned as is.
-      - If the path is relative, it is made absolute by appending it to [base_path]. *)
-  val to_absolute_path : t -> base_path:Absolute_path.t -> Absolute_path.t
+(** This module is re-exported as part of the [Fpath] module. For example:
+    [Fpath.classify]. *)
+module Export : sig
+  val classify
+    :  Fpath.t
+    -> [ `Absolute of Absolute_path.t | `Relative of Relative_path.t ]
 end